fix(theme): refractored the makeOriginal after we removed all seperate components makeThemes.
